main_code/requirements.txt
fastapi
uvicorn
google-genai
google-adk==1.5.0
httpx
a2a-sdk==0.2.10
litellm
python-dotenv
fastmcp==2.2.5
asyncclick
pytest
click
nest_asyncio